requirements.txt
gitdb==4.0.11
GitPython==3.1.41
pytz==2024.1
smmap==5.0.1
tqdm==4.66.4
